Day15/requirements.txt
fastapi
uvicorn
langchain
langchain-google-genai
faiss-cpu
python-dotenv
